Proof of concept, it's messy but working - read for more desc
So Just to check that the sorting in the content resolver were working I did this little mess to check if it was a viable solution. Ended up with the idea that it CAN work, but I will need to refactor everything from this commit in order to make this technique working correctly and fast enought. Maybe I will change the db to something with a faster read speed compared to SQLite. Also the access to the database doesn't fit my standart in term of coding and threading ! So I will have to work a lot to have something correct here.